operation overflow protection

+++ rosen/modules/render_service_base/include/animation/rs_render_spring_animation.h
@@ -30,7 +30,7 @@ public:
         const std::shared_ptr<RSRenderPropertyBase>& endValue);
 
     void DumpAnimationInfo(std::string& out) const override;
-    void SetSpringParameters(float response, float dampingRatio, float blendDuration = 0.0f);
+    void SetSpringParameters(float response, float dampingRatio, double blendDuration = 0.0f);
     void SetZeroThreshold(float zeroThreshold);
     void SetInitialVelocity(const std::shared_ptr<RSRenderPropertyBase>& velocity);
     void InheritSpringAnimation(const std::shared_ptr<RSRenderAnimation>& prevAnimation);
